# neverland/internal/docs/static/openapi.yaml
openapi: 3.1.0
info:
  title: Neverland
  description: Tinkerbell management gateway — REST API for bare-metal provisioning
  version: 0.1.0
servers:
  - url: http://10.11.30.91:8091
    description: LAN (k3s node)
paths:
...
  /api/v1/artifacts:
    get:
      summary: List OS image artifacts
      operationId: listArtifacts
      tags: [Artifacts]
      responses:
        "200":
          description: List of artifact files
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/ArtifactInfo"
    post:
      summary: Upload an OS image
      operationId: uploadArtifact
      tags: [Artifacts]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: Artifact uploaded
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                  size:
                    type: integer
                  url:
                    type: string
...
  /api/v1/artifacts/{filename}:
    parameters:
      - name: filename
        in: path
        required: true
        schema:
          type: string
    get:
      summary: Download an artifact (proxied from nginx)
      operationId: downloadArtifact
      tags: [Artifacts]
      responses:
        "200":
          description: File content
    delete:
      summary: Delete an artifact
      operationId: deleteArtifact
      tags: [Artifacts]
      responses:
        "200":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/beacon:
    get:
      summary: iPXE-format boot beacon (used by imgfetch)
      description: |
        Called by the iPXE menu via imgfetch. Query parameters carry the
        fingerprint. Response is itself an iPXE script that sets
        kontango_session_id, kontango_short_code, kontango_skip_claim.
      tags: [Boot]
      parameters:
        - in: query
          name: ipxe
          required: true
          schema: { type: integer, enum: [1] }
        - in: query
          name: mac
          required: true
          schema: { type: string }
        - in: query
          name: ip
          schema: { type: string }
        - in: query
          name: arch
          schema: { type: string, enum: [amd64, arm64] }
        - in: query
          name: platform
          schema: { type: string, enum: [pcbios, efi] }
        - in: query
          name: userclass
          schema: { type: string }
      responses:
        "200":
          description: iPXE script with set commands
          content:
            text/plain:
              schema: { type: string }
    post:
      summary: JSON boot beacon (used by Hook OS)
      tags: [Boot]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [level, fingerprint]
              properties:
                level:
                  type: string
                  enum: [ipxe, hook]
                session_id:
                  type: string
                  format: uuid
                fingerprint:
                  type: object
                  required: [mac]
                  properties:
                    mac:        { type: string }
                    ip:         { type: string }
                    arch:       { type: string }
                    platform:   { type: string }
                    userclass:  { type: string }
                    hostname:   { type: string }
                    dmi_serial: { type: string }
                    dmi_uuid:   { type: string }
                    cpu_id:     { type: string }
                    memory_kb:  { type: integer }
                    disk_count: { type: integer }
                    tpm_present: { type: boolean }
                    tpm_ek_pubkey: { type: string }
      responses:
        "200":
          description: Beacon receipt
          content:
            application/json:
              schema:
                type: object
                properties:
                  session_id: { type: string, format: uuid }
                  short_code: { type: string }
                  claim_url:  { type: string, format: uri }
                  skip_claim: { type: boolean }
                  watch_url:  { type: string, format: uri }
...
components:
...
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
...
    ArtifactInfo:
      type: object
      properties:
        name:
          type: string
          example: ubuntu-24.04-server-cloudimg-amd64.raw.gz
        size:
          type: integer
          example: 1234567890
        modified:
          type: string
          format: date-time
        url:
          type: string
          example: http://10.11.30.91:8080/ubuntu-24.04-server-cloudimg-amd64.raw.gz
...